+++ app/include_upgrade_testnet.go
@@ -16,6 +16,7 @@ import (
 	v41 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v4_1"
 	v42 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v4_2"
 	v43 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v4_3"
+	v44 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v4_4"
 	v4rc3 "github.com/babylonlabs-io/babylon/v4/app/upgrades/v4rc3/testnet"
 )
 
@@ -23,6 +24,7 @@ import (
 // it is also used for e2e testing
 func init() {
 	Upgrades = []upgrades.Upgrade{
+		v44.Upgrade,
 		v43.Upgrade,
 		v42.Upgrade,
 		v41.Upgrade,